Create search button on homepage and destination_location value carries over to flight_search page, move user pages to user folder, fix fetching of user details from db to the user dashboard, modify user routers, usersapi, and users_controller.
