+++ frontend/src/components/AdminFooter.jsx
@@ -1,4 +1,4 @@
-import logoImage from "../assets/logo-no-text.png"
+import logoImage from "../assets/logo-no-text.png";
 
 const AdminFooter = () => {
   return (
@@ -15,7 +15,7 @@ const AdminFooter = () => {
           />
         </figure>
         <p className="footer text-[11px]">
-          Discover the Skies with Nimbus Airways: Elevating Your Journey Beyond
+          Discover the Skies with NimbusAirways: Elevating Your Journey Beyond
           Boundaries
           <br />
           &copy; NimbusAirways 2024. All rights reserved.
