+++ frontend/src/layouts/router/UserDashboardRoutes.jsx
@@ -1,19 +1,21 @@
 import { Routes, Route } from "react-router-dom";
 import UserDashboardLayout from "../UserDashboardLayout";
-import UserDashboard from "../../pages/Dashboard/UserDashboard";
-import Bookings from "../../pages/Bookings";
-import FlightsSearchComponent from "@/pages/FlightsSearch";
+import UserDashboard from "../../pages/user/Dashboard/UserDashboard";
+import Bookings from "../../pages/user/Bookings";
+import FlightsSearchComponent from "../../pages/user/FlightsSearch";
 // eslint-disable-next-line react/prop-types
-const UserDashboardRoutes = ({addAlert}) => {
 
-
+const UserDashboardRoutes = ({ addAlert }) => {
   return (
     <>
       <UserDashboardLayout>
         <Routes>
-          <Route path="/" element={<UserDashboard />} />
+          <Route path="my_dashboard" element={<UserDashboard />} />
           <Route path="bookings" element={<Bookings addAlert={addAlert} />} />
-          <Route path="flight_search" element={<FlightsSearchComponent addAlert={addAlert} />} />
+          <Route
+            path="flight_search"
+            element={<FlightsSearchComponent addAlert={addAlert} />}
+          />
         </Routes>
       </UserDashboardLayout>
     </>
