+++ frontend/src/pages/home/home.jsx
@@ -1,22 +1,59 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import Footer from "../home/footer";
 import logoImage from "../../assets/logo.png";
 import homepageVideo from "../../assets/HomepageVideo.mp4";
 import { Link, useNavigate } from "react-router-dom";
+import FlightSearchDestination from "../../components/flightsearchdestination";
+import { indexedRoutesApi } from "../../lib/flightsapi";
 
 const Home = () => {
   const navigate = useNavigate();
+  const [destinationOptions, setDestinationOptions] = useState([]);
+  const [destination_location, setDestination_location] = useState("");
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const routesData = await indexedRoutesApi();
+        const uniqueDestinationLocations = Array.from(
+          new Set(routesData.map((route) => route.destination_location))
+        );
+        setDestinationOptions(uniqueDestinationLocations);
+      } catch (error) {
+        console.error("Error fetching initial flight information:", error);
+        addAlert("error", "Error fetching flights. Please try again.");
+      }
+    };
+    fetchData();
+  }, []);
+
+  const handleDestinationSelect = (selectedOption) => {
+    setDestination_location(selectedOption);
+  };
+
+  const handleSearch = async (event) => {
+    event.preventDefault();
+
+    if (!destination_location) {
+      addAlert("error", "Please choose a destination before searching.");
+      return;
+    }
+
+    localStorage.setItem("destination_from_homepage", destination_location);
+
+    navigate(`/flight_search`);
+  };
 
   return (
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "column",
-        height: "100vh",
-        overflow: "hidden",
-      }}
-    >
-      <div>
+    <>
+      <div
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          height: "100vh",
+          overflow: "hidden",
+        }}
+      >
         <div className="navbar bg-accent top-0">
           <div className="flex-1">
             <figure>
@@ -29,68 +66,68 @@ const Home = () => {
               />
             </figure>
           </div>
-          <div className="flex-none gap-2">
-            <div className="form-control">
-              <input
-                type="text"
-                placeholder="Enter a city"
-                className="input input-bordered bg-white text-sm w-20 md:w-auto"
-              />
+          <label className="form-control w-full max-w-xs mr-2 items-end">
+            <FlightSearchDestination
+              destinationOptions={destinationOptions}
+              onSelect={handleDestinationSelect}
+              setDestination_location={setDestination_location}
+            />
+          </label>
+          <div>
+            <button className="btn btn-info mr-2" onClick={handleSearch}>
+              Search
+            </button>
+          </div>
+          <div className="dropdown dropdown-end">
+            <div tabIndex={0} role="button" className="btn btn-primary">
+              <div className="w-10 rounded-full text-black">Menu</div>
             </div>
-            <div>
-              <button className="btn btn-info">Search</button>
-            </div>
-            <div className="dropdown dropdown-end">
-              <div tabIndex={0} role="button" className="btn btn-primary ">
-                <div className="w-10 rounded-full text-white">Menu</div>
-              </div>
-              <ul
-                tabIndex={0}
-                className="mt-3 z-[1] p-2 shadow menu menu-sm dropdown-content bg-base-100 rounded-box w-52"
-              >
-                <li>
-                  <a
-                    className="justify-between"
-                    onClick={() => navigate("/login")}
-                  >
-                    Login/Register
-                  </a>
-                </li>
-                <li>
-                  <Link to="/about_us">About Us</Link>
-                </li>
-                <li>
-                  <Link to="/contact_us">Contact Us</Link>
-                </li>
-              </ul>
-            </div>
+            <ul
+              tabIndex={0}
+              className="mt-3 z-[1] p-2 shadow menu menu-sm dropdown-content bg-base-100 rounded-box w-52"
+            >
+              <li>
+                <a
+                  className="justify-between"
+                  onClick={() => navigate("/login")}
+                >
+                  Login/Register
+                </a>
+              </li>
+              <li>
+                <Link to="/about_us">About Us</Link>
+              </li>
+              <li>
+                <Link to="/contact_us">Contact Us</Link>
+              </li>
+            </ul>
           </div>
         </div>
-      </div>
 
-      <div
-        style={{
-          position: "relative",
-          flex: 1,
-          height: "5%",
-          margin: "0",
-        }}
-      >
-        <video
-          src={homepageVideo}
-          autoPlay
-          loop
-          muted
+        <div
           style={{
-            width: "100%",
-            height: "100%",
-            objectFit: "cover",
+            position: "relative",
+            flex: 1,
+            height: "5%",
+            margin: "0",
           }}
-        />
-      </div>
+        >
+          <video
+            src={homepageVideo}
+            autoPlay
+            loop
+            muted
+            style={{
+              width: "100%",
+              height: "100%",
+              objectFit: "cover",
+            }}
+          />
+        </div>
 
-      <Footer />
-    </div>
+        <Footer />
+      </div>
+    </>
   );
 };
